requirements.txt
# requirements
numpy
scikit-learn==0.22.1
tensorflow==2.1.0
joblib==0.14.1
pillow=7.0.0
